moving upload function to utils
