+++ sar_pipeline/nci/preparation/geoid.py
@@ -12,8 +12,8 @@ import rasterio
 import rasterio.transform
 import shapely.geometry
 from rasterio.crs import CRS
-from ...utils.raster import read_raster_with_bounds
-from ...utils.rio_tools import translate_profile, reproject_arr_to_match_profile
+from sar_pipeline.utils.raster import read_raster_with_bounds
+from sar_pipeline.utils.rio_tools import translate_profile, reproject_arr_to_match_profile
 
 
 def read_geoid(
